TIVariables now store the variable name as a string; identity is based on this name rather than the typeinfo itself.
variance barriers added. since the types that appear in ascriptions and local type definitions are invisble outside that scope, variables from outside scopes can be used in an unrestricted manner. therefore, we place a variance barrier at the below each code block, which indicates that any variable bound at that scope or lower should be interpreted as invariant.